requirements.txt
streamlit==1.28.0
openai==0.28.0
python-dotenv==1.0.0
googletrans==4.0.0-rc1
plotly==5.15.0
pandas==2.0.3
requests==2.31.0